requirements.txt
attr==0.3.1
bbutils==0.2.0.0
colorama==0.4.4
coverage==5.3
XlsxWriter==1.3.7
yattag==1.14.0
